+++ api/src/main/kotlin/com/backgu/amaker/auth/dto/oauth/google/GoogleOAuth2AccessTokenDto.kt
@@ -1,4 +1,4 @@
-package com.backgu.amaker.auth.dto
+package com.backgu.amaker.auth.dto.oauth.google
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
